requirements/base.txt
# Bleeding edge Django
Django==1.7.4

# Configuration
django-configurations==0.8
django-secure==1.0.1
django-cache-url==0.8.0
dj-database-url==0.3.0

# Forms
django-braces==1.4.0
django-crispy-forms==1.4.0
django-floppyforms==1.3.0

# Models
django-model-utils==2.2

# Templates
jingo==0.7

# Images
Pillow==2.7.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.19.0

# For the persistence stores
psycopg2==2.5.4

# Unicode slugification
# unicode-slugify==0.1.3
# django-autoslug==1.7.2

# Useful things
django-avatar==2.0

# Time zones support
pytz==2014.10

# Your custom requirements go here
django-durationfield==0.5.1
django-betterforms==1.1.1
arrow==0.5.0
